Re-write the directory/dotfile linking
Instead of the weird hack using `dot.` prefix to decide what to link we
use `$NFA/home/*` with the actual names the same. Stops the need for any
messy name mapping.  Fixes issue with what to do with directories or
files that don't start with dot.
Uses the new nfa logging and fixed dry running.